+++ src/main/java/tn/esprit/entities/TransportLocal.java
@@ -1,6 +1,5 @@
 package tn.esprit.entities;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class TransportLocal {
@@ -11,7 +10,7 @@ public class TransportLocal {
     private String paysArrivee;
     private LocalDate dateDepart;
     private LocalDate dateRetour;
-    private BigDecimal prix;
+    private double prix;
     private int idVoyage;
     private int nbrPlaces;
 
@@ -23,13 +22,13 @@ public class TransportLocal {
 
     public TransportLocal(int idTransport, String compagnie, TypeTransport typeTransport,
                           String paysDepart, String paysArrivee, LocalDate dateDepart, LocalDate dateRetour,
-                          BigDecimal prix, int idVoyage) {
+                          double prix, int idVoyage) {
         this(idTransport, compagnie, typeTransport, paysDepart, paysArrivee, dateDepart, dateRetour, prix, idVoyage, 10);
     }
 
     public TransportLocal(int idTransport, String compagnie, TypeTransport typeTransport,
                           String paysDepart, String paysArrivee, LocalDate dateDepart, LocalDate dateRetour,
-                          BigDecimal prix, int idVoyage, int nbrPlaces) {
+                          double prix, int idVoyage, int nbrPlaces) {
         this.idTransport = idTransport;
         this.compagnie = compagnie;
         this.typeTransport = typeTransport;
@@ -56,8 +55,8 @@ public class TransportLocal {
     public void setDateDepart(LocalDate dateDepart) { this.dateDepart = dateDepart; }
     public LocalDate getDateRetour() { return dateRetour; }
     public void setDateRetour(LocalDate dateRetour) { this.dateRetour = dateRetour; }
-    public BigDecimal getPrix() { return prix; }
-    public void setPrix(BigDecimal prix) { this.prix = prix; }
+    public double getPrix() { return prix; }
+    public void setPrix(double prix) { this.prix = prix; }
     public int getIdVoyage() { return idVoyage; }
     public void setIdVoyage(int idVoyage) { this.idVoyage = idVoyage; }
     public int getNbrPlaces() { return nbrPlaces; }
